requirements.txt
torch>=2.1.0
torchvision>=0.16.0
compressai>=1.2.6
pytorch-msssim>=1.0.0
numpy>=1.26.4,<2.3
pillow>=10.4.0
tqdm>=4.66.4
wandb>=0.17.0
